Create Double Linked list
his repository contains an implementation of a Doubly Linked List in C. A doubly linked list is a linear data structure where each node contains three parts: 

Data: The value of the node. 

Previous Pointer: A pointer to the previous node in the list. 

Next Pointer: A pointer to the next node in the list. 

Unlike singly linked lists, where traversal is only possible in one direction (forward), doubly linked lists allow for bidirectional traversal (both forward and backward). This additional flexibility comes with a slight increase in memory usage due to the extra pointer (prev) in each node.